refactor: fetching user role
