+++ src/app/layout.tsx
@@ -2,6 +2,7 @@ import type { Metadata } from "next";
 import "./globals.css";
 import { ClerkProvider } from "@clerk/nextjs";
 import Navbar from "@/components/Navbar";
+import { UserRoleProvider } from "@/components/UserRoleProvider";
 
 //! Update metadata to match your project
 export const metadata: Metadata = {
@@ -15,14 +16,16 @@ export const metadata: Metadata = {
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <ClerkProvider>
-      <html lang="en" className="h-full">
-        <body className="flex flex-col h-full bg-light-gray">
-          <header>
-            <Navbar />
-          </header>
-          {children}
-        </body>
-      </html>
+      <UserRoleProvider>
+        <html lang="en" className="h-full">
+          <body className="flex flex-col h-full bg-light-gray">
+            <header>
+              <Navbar />
+            </header>
+            {children}
+          </body>
+        </html>
+      </UserRoleProvider>
     </ClerkProvider>
   );
 }
